+++ src/Aspire.Hosting.Azure.PostgreSQL/README.md
@@ -48,21 +48,10 @@ var myService = builder.AddProject<Projects.MyService>()
                        .WithReference(postgresdb);
 ```
 
-The `WithReference` method configures a connection in the `MyService` project named `postgresdb`. By default, `AddAzurePostgresFlexibleServer` configures [Microsoft Entra ID](https://learn.microsoft.com/azure/postgresql/flexible-server/concepts-azure-ad-authentication) authentication. This requires changes to applications that need to connect to these resources. In the _Program.cs_ file of `MyService`, the database connection can be consumed using the client library [Aspire.Npgsql](https://www.nuget.org/packages/Aspire.Npgsql) and [Azure.Identity](https://www.nuget.org/packages/Azure.Identity):
+The `WithReference` method configures a connection in the `MyService` project named `postgresdb`. By default, `AddAzurePostgresFlexibleServer` configures [Microsoft Entra ID](https://learn.microsoft.com/azure/postgresql/flexible-server/concepts-azure-ad-authentication) authentication. This requires changes to applications that need to connect to these resources. In the _Program.cs_ file of `MyService`, the database connection can be consumed using the client library [Aspire.Azure.Npgsql](https://www.nuget.org/packages/Aspire.Azure.Npgsql) or [Aspire.Azure.Npgsql.EntityFrameworkCore.PostgreSQL](https://www.nuget.org/packages/Aspire.Azure.Npgsql.EntityFrameworkCore.PostgreSQL):
 
 ```csharp
-builder.AddNpgsqlDataSource("postgresdb", configureDataSourceBuilder: dataSourceBuilder =>
-{
-    if (string.IsNullOrEmpty(dataSourceBuilder.ConnectionStringBuilder.Password))
-    {
-        dataSourceBuilder.UsePeriodicPasswordProvider(async (_, ct) =>
-        {
-            var credentials = new DefaultAzureCredential();
-            var token = await credentials.GetTokenAsync(new TokenRequestContext(["https://ossrdbms-aad.database.windows.net/.default"]), ct);
-            return token.Token;
-        }, TimeSpan.FromHours(24), TimeSpan.FromSeconds(10));
-    }
-});
+builder.AddAzureNpgsqlDataSource("postgresdb");
 ```
 
 ## Additional documentation
