+++ src/Components/Aspire.Azure.Npgsql.EntityFrameworkCore.PostgreSQL/AspireAzureEFPostgreSqlExtensions.cs
@@ -3,10 +3,9 @@
 
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
-using System.Text.Json;
+using Aspire;
 using Aspire.Azure.Npgsql.EntityFrameworkCore.PostgreSQL;
 using Aspire.Npgsql.EntityFrameworkCore.PostgreSQL;
-using Azure.Core;
 using Azure.Identity;
 using Microsoft.EntityFrameworkCore;
 using Npgsql;
@@ -14,8 +13,6 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Infrastructure.Internal;
 
 #if NET9_0_OR_GREATER
 using Microsoft.Extensions.DependencyInjection;
-#else
-using Aspire;
 #endif
 
 namespace Microsoft.Extensions.Hosting;
@@ -26,11 +23,6 @@ namespace Microsoft.Extensions.Hosting;
 public static partial class AspireAzureEFPostgreSqlExtensions
 {
     private const DynamicallyAccessedMemberTypes RequiredByEF = DynamicallyAccessedMemberTypes.PublicConstructors | DynamicallyAccessedMemberTypes.NonPublicConstructors | DynamicallyAccessedMemberTypes.PublicProperties;
-    private const string AzureDatabaseForPostgresSqlScope = "https://ossrdbms-aad.database.windows.net/.default";
-    private const string AzureManagementScope = "https://management.azure.com/.default";
-
-    private static readonly TokenRequestContext s_databaseForPostgresSqlTokenRequestContext = new([AzureDatabaseForPostgresSqlScope]);
-    private static readonly TokenRequestContext s_managementTokenRequestContext = new([AzureManagementScope]);
 
     /// <summary>
     /// Registers the given <see cref="DbContext" /> as a service in the services provided by the <paramref name="builder"/>.
@@ -129,7 +121,7 @@ public static partial class AspireAzureEFPostgreSqlExtensions
 
     private static void ConfigureDbContextOptionsBuilder(AzureNpgsqlEntityFrameworkCorePostgreSQLSettings settings, DbContextOptionsBuilder dbContextOptionsBuilder)
     {
-        // The connection string required the username to be provided. Since it will depend on the Managed Identity that is used
+        // The connection string requires the username to be provided. Since it will depend on the Managed Identity that is used
         // we attempt to get the username from the access token.
 
         var credential = settings.Credential ?? new DefaultAzureCredential();
@@ -146,18 +138,18 @@ public static partial class AspireAzureEFPostgreSqlExtensions
         if (string.IsNullOrEmpty(dataSourceBuilder.ConnectionStringBuilder.Username))
         {
             // Ensure to use the management scope, so the token contains user names for all managed identity types - e.g. user and service principal
-            var token = credential.GetToken(s_managementTokenRequestContext, default);
+            var token = credential.GetToken(ManagedIdentityTokenCredentialHelpers.ManagementTokenRequestContext, default);
 
-            if (TryGetUsernameFromToken(token.Token, out var username))
+            if (ManagedIdentityTokenCredentialHelpers.TryGetUsernameFromToken(token.Token, out var username))
             {
                 dataSourceBuilder.ConnectionStringBuilder.Username = username;
             }
             else
             {
                 // Otherwise check using the PostgresSql scope
-                token = credential.GetToken(s_databaseForPostgresSqlTokenRequestContext, default);
+                token = credential.GetToken(ManagedIdentityTokenCredentialHelpers.DatabaseForPostgresSqlTokenRequestContext, default);
 
-                if (TryGetUsernameFromToken(token.Token, out username))
+                if (ManagedIdentityTokenCredentialHelpers.TryGetUsernameFromToken(token.Token, out username))
                 {
                     dataSourceBuilder.ConnectionStringBuilder.Username = username;
                 }
@@ -169,86 +161,11 @@ public static partial class AspireAzureEFPostgreSqlExtensions
         if (string.IsNullOrEmpty(dataSourceBuilder.ConnectionStringBuilder.Password))
         {
             dataSourceBuilder.UsePasswordProvider(
-                passwordProvider: _ => credential.GetToken(s_databaseForPostgresSqlTokenRequestContext, default).Token,
-                passwordProviderAsync: async (_, ct) => (await credential.GetTokenAsync(s_databaseForPostgresSqlTokenRequestContext, default).ConfigureAwait(false)).Token
+                passwordProvider: _ => credential.GetToken(ManagedIdentityTokenCredentialHelpers.DatabaseForPostgresSqlTokenRequestContext, default).Token,
+                passwordProviderAsync: async (_, ct) => (await credential.GetTokenAsync(ManagedIdentityTokenCredentialHelpers.DatabaseForPostgresSqlTokenRequestContext, default).ConfigureAwait(false)).Token
             );
 
             dbContextOptionsBuilder.UseNpgsql(dataSourceBuilder.Build());
         }
     }
-
-    private static bool TryGetUsernameFromToken(string jwtToken, out string? username)
-    {
-        username = null;
-
-        // Split the token into its parts (Header, Payload, Signature)
-        var tokenParts = jwtToken.Split('.');
-        if (tokenParts.Length != 3)
-        {
-            return false;
-        }
-
-        // The payload is the second part, Base64Url encoded
-        var payload = tokenParts[1];
-
-        // Add padding if necessary
-        payload = AddBase64Padding(payload);
-
-        // Decode the payload from Base64Url
-        var decodedBytes = Convert.FromBase64String(payload);
-
-        // Parse the decoded payload as JSON
-        var reader = new Utf8JsonReader(decodedBytes);
-        var payloadJson = JsonElement.ParseValue(ref reader);
-
-        // Try to get the username from 'xms_mirid', 'upn', 'preferred_username', or 'unique_name' claims
-        if (payloadJson.TryGetProperty("xms_mirid", out var xms_mirid) &&
-            xms_mirid.GetString() is string xms_miridString &&
-            ParsePrincipalName(xms_miridString) is string principalName)
-        {
-            username = principalName;
-        }
-        else if (payloadJson.TryGetProperty("upn", out var upn))
-        {
-            username = upn.GetString();
-        }
-        else if (payloadJson.TryGetProperty("preferred_username", out var preferredUsername))
-        {
-            username = preferredUsername.GetString();
-        }
-        else if (payloadJson.TryGetProperty("unique_name", out var uniqueName))
-        {
-            username = uniqueName.GetString();
-        }
-
-        return username != null;
-    }
-
-    // parse the xms_mirid claim which look like
-    // /subscriptions/{subId}/resourcegroups/{resourceGroup}/providers/Microsoft.ManagedIdentity/userAssignedIdentities/{principalName}
-    private static string? ParsePrincipalName(string xms_mirid)
-    {
-        var lastSlashIndex = xms_mirid.LastIndexOf('/');
-        if (lastSlashIndex == -1)
-        {
-            return null;
-        }
-
-        var beginning = xms_mirid.AsSpan(0, lastSlashIndex);
-        var principalName = xms_mirid.AsSpan(lastSlashIndex + 1);
-
-        if (principalName.IsEmpty || !beginning.EndsWith("providers/Microsoft.ManagedIdentity/userAssignedIdentities", StringComparison.OrdinalIgnoreCase))
-        {
-            return null;
-        }
-
-        return principalName.ToString();
-    }
-
-    private static string AddBase64Padding(string base64) => (base64.Length % 4) switch
-    {
-        2 => base64 + "==",
-        3 => base64 + "=",
-        _ => base64,
-    };
 }
