+++ src/Components/Aspire.Azure.Npgsql/Aspire.Azure.Npgsql.csproj
@@ -10,6 +10,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Include="..\Common\ManagedIdentityTokenCredentialHelpers.cs" Link="ManagedIdentityTokenCredentialHelpers.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Azure" />
     <ProjectReference Include="..\Aspire.Npgsql\Aspire.Npgsql.csproj" />
   </ItemGroup>
