+++ src/Components/Aspire.Azure.Npgsql/AspireAzureNpgsqlExtensions.cs
@@ -2,10 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Diagnostics;
-using System.Text.Json;
+using Aspire;
 using Aspire.Azure.Npgsql;
 using Aspire.Npgsql;
-using Azure.Core;
 using Azure.Identity;
 using Microsoft.Extensions.DependencyInjection;
 using Npgsql;
@@ -17,12 +16,6 @@ namespace Microsoft.Extensions.Hosting;
 /// </summary>
 public static class AspireAzureNpgsqlExtensions
 {
-    private const string AzureDatabaseForPostgresSqlScope = "https://ossrdbms-aad.database.windows.net/.default";
-    private const string AzureManagementScope = "https://management.azure.com/.default";
-
-    private static readonly TokenRequestContext s_databaseForPostgresSqlTokenRequestContext = new([AzureDatabaseForPostgresSqlScope]);
-    private static readonly TokenRequestContext s_managementTokenRequestContext = new([AzureManagementScope]);
-
     /// <summary>
     /// Registers <see cref="NpgsqlDataSource"/> service for connecting PostgreSQL database with Npgsql client.
     /// Configures health check, logging and telemetry for the Npgsql client.
@@ -106,7 +99,7 @@ public static class AspireAzureNpgsqlExtensions
 
     private static void ConfigureDataSourceBuilder(AzureNpgsqlSettings settings, NpgsqlDataSourceBuilder dataSourceBuilder)
     {
-        // The connection string required the username to be provided. Since it will depend on the Managed Identity that is used
+        // The connection string requires the username to be provided. Since it will depend on the Managed Identity that is used
         // we attempt to get the username from the access token.
 
         var credential = settings.Credential ?? new DefaultAzureCredential();
@@ -114,18 +107,18 @@ public static class AspireAzureNpgsqlExtensions
         if (string.IsNullOrEmpty(dataSourceBuilder.ConnectionStringBuilder.Username))
         {
             // Ensure to use the management scope, so the token contains user names for all managed identity types - e.g. user and service principal
-            var token = credential.GetToken(s_managementTokenRequestContext, default);
+            var token = credential.GetToken(ManagedIdentityTokenCredentialHelpers.ManagementTokenRequestContext, default);
 
-            if (TryGetUsernameFromToken(token.Token, out var username))
+            if (ManagedIdentityTokenCredentialHelpers.TryGetUsernameFromToken(token.Token, out var username))
             {
-                 dataSourceBuilder.ConnectionStringBuilder.Username = username;
+                dataSourceBuilder.ConnectionStringBuilder.Username = username;
             }
             else
             {
                 // Otherwise check using the PostgresSql scope
-                token = credential.GetToken(s_databaseForPostgresSqlTokenRequestContext, default);
+                token = credential.GetToken(ManagedIdentityTokenCredentialHelpers.DatabaseForPostgresSqlTokenRequestContext, default);
 
-                if (TryGetUsernameFromToken(token.Token, out username))
+                if (ManagedIdentityTokenCredentialHelpers.TryGetUsernameFromToken(token.Token, out username))
                 {
                     dataSourceBuilder.ConnectionStringBuilder.Username = username;
                 }
@@ -140,84 +133,9 @@ public static class AspireAzureNpgsqlExtensions
             // The token is not cached since it is refreshed for each new physical connection, or when it has expired.
 
             dataSourceBuilder.UsePasswordProvider(
-                passwordProvider: _ => credential.GetToken(s_databaseForPostgresSqlTokenRequestContext, default).Token,
-                passwordProviderAsync: async (_, ct) => (await credential.GetTokenAsync(s_databaseForPostgresSqlTokenRequestContext, default).ConfigureAwait(false)).Token
+                passwordProvider: _ => credential.GetToken(ManagedIdentityTokenCredentialHelpers.DatabaseForPostgresSqlTokenRequestContext, default).Token,
+                passwordProviderAsync: async (_, ct) => (await credential.GetTokenAsync(ManagedIdentityTokenCredentialHelpers.DatabaseForPostgresSqlTokenRequestContext, default).ConfigureAwait(false)).Token
             );
         }
     }
-
-    private static bool TryGetUsernameFromToken(string jwtToken, out string? username)
-    {
-        username = null;
-
-        // Split the token into its parts (Header, Payload, Signature)
-        var tokenParts = jwtToken.Split('.');
-        if (tokenParts.Length != 3)
-        {
-            return false;
-        }
-
-        // The payload is the second part, Base64Url encoded
-        var payload = tokenParts[1];
-
-        // Add padding if necessary
-        payload = AddBase64Padding(payload);
-
-        // Decode the payload from Base64Url
-        var decodedBytes = Convert.FromBase64String(payload);
-
-        // Parse the decoded payload as JSON
-        var reader = new Utf8JsonReader(decodedBytes);
-        var payloadJson = JsonElement.ParseValue(ref reader);
-
-        // Try to get the username from 'xms_mirid', 'upn', 'preferred_username', or 'unique_name' claims
-        if (payloadJson.TryGetProperty("xms_mirid", out var xms_mirid) &&
-            xms_mirid.GetString() is string xms_miridString &&
-            ParsePrincipalName(xms_miridString) is string principalName)
-        {
-            username = principalName;
-        }
-        else if (payloadJson.TryGetProperty("upn", out var upn))
-        {
-            username = upn.GetString();
-        }
-        else if (payloadJson.TryGetProperty("preferred_username", out var preferredUsername))
-        {
-            username = preferredUsername.GetString();
-        }
-        else if (payloadJson.TryGetProperty("unique_name", out var uniqueName))
-        {
-            username = uniqueName.GetString();
-        }
-
-        return username != null;
-    }
-
-    // parse the xms_mirid claim which look like
-    // /subscriptions/{subId}/resourcegroups/{resourceGroup}/providers/Microsoft.ManagedIdentity/userAssignedIdentities/{principalName}
-    private static string? ParsePrincipalName(string xms_mirid)
-    {
-        var lastSlashIndex = xms_mirid.LastIndexOf('/');
-        if (lastSlashIndex == -1)
-        {
-            return null;
-        }
-
-        var beginning = xms_mirid.AsSpan(0, lastSlashIndex);
-        var principalName = xms_mirid.AsSpan(lastSlashIndex + 1);
-
-        if (principalName.IsEmpty || !beginning.EndsWith("providers/Microsoft.ManagedIdentity/userAssignedIdentities", StringComparison.OrdinalIgnoreCase))
-        {
-            return null;
-        }
-
-        return principalName.ToString();
-    }
-
-    private static string AddBase64Padding(string base64) => (base64.Length % 4) switch
-    {
-        2 => base64 + "==",
-        3 => base64 + "=",
-        _ => base64,
-    };
 }
